+++ packages/skpm/package.json
@@ -38,7 +38,7 @@
   },
   "homepage": "https://github.com/skpm/skpm#readme",
   "dependencies": {
-    "@skpm/utils": "^0.1.6",
+    "@skpm/internal-utils": "^0.1.6",
     "bluebird": "^3.5.1",
     "chalk": "^1.1.3",
     "cross-spawn-promise": "^0.10.1",
